geotiff tests: consolidate unit cluster + slim conftest (#2414)
PR 11 of the GeoTIFF test consolidation epic (#2390).

* Create xrspatial/geotiff/tests/unit/ for pure-unit helper coverage:
  - unit/test_header.py: relocated from test_header.py
  - unit/test_geotags.py: relocated from test_geotags.py
  - unit/test_safe_xml.py: relocated from test_gdal_metadata_xml_escape_1614.py
  - unit/test_compression.py: merged from test_packbits_jit_2048.py
    and test_packbits_jit_2049.py
* Drop the pytest_collection_modifyitems socketserver hook in
  conftest.py. Every HTTP/loopback test now carries @requires_loopback
  explicitly (added to the parity matrix, sidecar, parallel-decode,
  remote-byte-order, and write/test_cog HTTP cases).
* Keep the helper / marker re-exports under __all__ so legacy
  conftest imports continue to resolve.
* Delete stale audit notes CLUSTER_AUDIT_PR5/8/9.md that leaked from
  earlier PRs.
* Update the release-gate rst row for writer.gdal_metadata_xml to
  cite the relocated unit/test_safe_xml.py.

CLUSTER_AUDIT_PR11.md ships in this commit and is removed in a
follow-up commit before merge.
