+++ xrspatial/geotiff/tests/conftest.py
@@ -1,12 +1,11 @@
 """Shared fixtures for geotiff tests.
 
-The TIFF builder, markers, and capability probes now live under
+The TIFF builder, markers, and capability probes live under
 ``_helpers/``. This module re-exports them so legacy imports such as
 ``from .conftest import make_minimal_tiff`` or
 ``from xrspatial.geotiff.tests.conftest import requires_gpu`` keep
-working. PR 11 of the consolidation epic (#2390) drops the
-``pytest_collection_modifyitems`` socketserver hack below; until then
-it stays in place to keep the existing loopback-skip behaviour.
+working without touching every test file. New tests should import the
+symbols directly from ``_helpers`` instead.
 """
 from __future__ import annotations
 
@@ -30,61 +29,6 @@ __all__ = [
     "requires_integration",
     "requires_loopback",
 ]
-
-
-def pytest_collection_modifyitems(config, items):
-    """Auto-skip tests that stand up a loopback HTTP server when the
-    sandbox denies socket bind.
-
-    A test needs loopback iff its function body or any fixture in its
-    closure references ``socketserver.TCPServer(`` or invokes the
-    file-local ``_serve(`` helper. This is finer-grained than skipping
-    every test in a module that imports ``socketserver``: mixed files
-    (e.g. ``test_miniswhite_backend_parity_1797.py`` has both HTTP and
-    a local-file GPU test) keep their non-HTTP coverage in restricted
-    sandboxes.
-    """
-    if loopback_available():
-        return
-
-    import inspect
-
-    def _source_of(obj) -> str:
-        try:
-            return inspect.getsource(obj)
-        except (OSError, TypeError):
-            return ''
-
-    def _references_loopback(src: str) -> bool:
-        return 'socketserver.TCPServer(' in src or '_serve(' in src
-
-    skip_marker = pytest.mark.skip(
-        reason="loopback bind unavailable in this environment"
-    )
-    for item in items:
-        needs_skip = False
-
-        func = getattr(item, 'function', None)
-        if func is not None and _references_loopback(_source_of(func)):
-            needs_skip = True
-
-        if not needs_skip:
-            fixtureinfo = getattr(item, '_fixtureinfo', None)
-            if fixtureinfo is not None:
-                name2defs = getattr(fixtureinfo, 'name2fixturedefs', {})
-                for fname in getattr(fixtureinfo, 'names_closure', ()):
-                    for fdef in name2defs.get(fname, ()):
-                        ffunc = getattr(fdef, 'func', None)
-                        if ffunc is not None and _references_loopback(
-                            _source_of(ffunc)
-                        ):
-                            needs_skip = True
-                            break
-                    if needs_skip:
-                        break
-
-        if needs_skip:
-            item.add_marker(skip_marker)
 
 
 @pytest.fixture
